Adds a search to the GPS UI, and typescriptifies it (#95323)
